% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nepal_wildlife.R
\docType{data}
\name{nepal_wildlife}
\alias{nepal_wildlife}
\title{Wildlife found in Nepal}
\format{
A data frame with 1409 rows and 4 variables:
\describe{
\item{Scientific Name}{Binomial nomenclature in \emph{genus} \emph{species} format}
\item{Common Name}{Commonly recognized English name}
\item{Family}{Family of the species}
\item{Taxonomic Group}{Taxonomic classification based on major groups such as mammals, birds and reptiles}
}
}
\source{
\url{https://mol.org/regions/?region=Nepal&regionid=2f3b667a-51b4-4aaa-be9b-005765b1be65&bounds=80.0601,26.3475,88.204,30.447}
}
\usage{
nepal_wildlife
}
\description{
A dataset containing common name, scientific name, family and taxonomic group of animals found in Nepal.
}
\keyword{datasets}
